[UPDATE] Change Count

+++ project09/src/main/java/com/example/project09/payload/post/response/PostResultResponse.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 public class PostResultResponse {
-    private Integer count;
+    private long count;
     private List<PostResponse> posts;
 }
